+++ client/src/features/viewers/lower-thirds/LowerClean.jsx
@@ -1,20 +1,21 @@
 import React, { useEffect, useState } from 'react';
 import { AnimatePresence, motion } from 'framer-motion';
 
-import NavLogo from "../../../../common/components/nav/NavLogo";
+import NavLogo from "../../../common/components/nav/NavLogo";
 
-import './LowerLines.scss';
+import './LowerClean.scss';
 
-export default function LowerLines(props) {
+export default function LowerClean(props) {
   const { lower, title, options } = props;
   const defaults = {
     size: 1,
     transitionIn: 3,
     textColour: '#fffffa',
-    bgColour: '#00000033',
+    posX: 50,
+    posY: 700,
   };
-  const [showLower, setShowLower] = useState(true);
 
+  const [showLower, setShowLower] = useState(true);
   // Unmount if fadeOut
   useEffect(() => {
     if (!options.fadeOut) return;
@@ -33,17 +34,12 @@ export default function LowerLines(props) {
     return () => clearTimeout(timeout);
   }, [options.fadeOut, options.transitionIn, defaults.transitionIn]);
 
-  useEffect(() => {
-    setShowLower(title.showNow);
-  }, [title.showNow]);
-
   // Format messages
   const showLowerMessage = lower.text !== '' && lower.visible;
 
   // motion
   // transition segments
   const t = options.transitionIn || defaults.transitionIn;
-  const eight = t / 8;
   const quarter = t / 4;
   const third = t / 3;
   const half = t / 2;
@@ -60,19 +56,6 @@ export default function LowerLines(props) {
     },
     exit: {
       opacity: 0,
-      x: -1000,
-      transition: {
-        duration: third,
-      },
-    },
-  };
-
-  const titleContainerVariants = {
-    hidden: {
-      left: -1000,
-    },
-    visible: {
-      left: 0,
       transition: {
         duration: third,
       },
@@ -92,37 +75,11 @@ export default function LowerLines(props) {
     },
   };
 
-  const subtitleContainerVariants = {
-    hidden: {
-      left: -1000,
-    },
-    visible: {
-      left: 0,
-      transition: {
-        delay: eight,
-        duration: third,
-      },
-    },
-  };
-
-  const subtitleVariants = {
-    hidden: {
-      opacity: 0,
-    },
-    visible: {
-      opacity: 1,
-      transition: {
-        delay: third,
-        duration: third * 2,
-      },
-    },
-  };
-
   const sizeMultiplier = (options.size || 1) * 4;
 
   return (
     <div
-      className='lower-third lines'
+      className='lower-third clean'
       style={{
         backgroundColor: options.keyColour || defaults.keyColour,
         color: options.textColour || defaults.textColour,
@@ -136,32 +93,21 @@ export default function LowerLines(props) {
         {showLower && (
           <motion.div
             className='lower-container'
-            style={{ backgroundColor: options.bgColour || defaults.bgColour }}
+            style={{
+              backgroundColor: options.bgColour || defaults.bgColour,
+              top: options.posY || defaults.posY,
+              left: options.posX || defaults.posX,
+            }}
             variants={lowerThirdVariants}
             initial='hidden'
             animate='visible'
             exit='exit'
           >
-            <motion.div
-              className='title-container'
-              variants={titleContainerVariants}
-            >
-              <motion.div className='title' variants={titleVariants}>
-                {title.titleNow}
-              </motion.div>
-              <div className='title-decor' />
+            <motion.div className='title' variants={titleVariants}>
+              {title.titleNow}
             </motion.div>
-            <motion.div
-              className='subtitle-container'
-              variants={subtitleContainerVariants}
-            >
-              <div className='sub-decor' />
-              <motion.div
-                className='subtitle'
-                variants={subtitleVariants}
-              >
-                {title.presenterNow}
-              </motion.div>
+            <motion.div className='subtitle' variants={titleVariants}>
+              {title.presenterNow}
             </motion.div>
           </motion.div>
         )}
@@ -171,6 +117,9 @@ export default function LowerLines(props) {
         {showLowerMessage && (
           <motion.div
             className='message-container'
+            style={{
+              backgroundColor: options.bgColour || defaults.bgColour,
+            }}
             key='modal'
             initial={{ opacity: 0 }}
             animate={{ opacity: 1 }}
