+++ client/src/features/viewers/lower-thirds/LowerClean.scss
