+++ client/src/features/viewers/lower-thirds/LowerLines.scss
